symbols: move `SpirvAttribute` to the attr module.
